+++ app/app.go
@@ -3,15 +3,43 @@ package app
 import (
 	"bytes"
 	"fmt"
+	"os"
 	"path"
 
 	abci "github.com/tendermint/abci/types"
 	"github.com/tendermint/go-wire"
-	"github.com/tendermint/merkleeyes/iavl"
+	tmflags "github.com/tendermint/tmlibs/cli/flags"
 	cmn "github.com/tendermint/tmlibs/common"
 	dbm "github.com/tendermint/tmlibs/db"
+	"github.com/tendermint/tmlibs/log"
 	"github.com/tendermint/tmlibs/merkle"
+
+	"github.com/tendermint/merkleeyes/iavl"
 )
+
+// TODO: this is here to avoid breaking changes.
+// the logger should go inside MerkleEyesApp as another arg, but that must
+// go into a minor release, not a quick patch.
+var (
+	defaultLogLevel = "error"
+	logger          = log.NewTMLogger(log.NewSyncWriter(os.Stdout)).With("module", "merkleeyes")
+)
+
+// SetLogger sets a brand new logger
+func SetLogger(newlog log.Logger) {
+	logger = newlog
+}
+
+// SetLogLevel sets the log_level on the package logger
+func SetLogLevel(level string) (err error) {
+	logger, err = tmflags.ParseLogLevel(level, logger, defaultLogLevel)
+	return
+}
+
+// SetTraceLogger adds error tracing to the logger
+func SetTraceLogger() {
+	logger = log.NewTracingLogger(logger)
+}
 
 // MerkleEyesApp is a Merkle KV-store served as an ABCI app
 type MerkleEyesApp struct {
@@ -72,13 +100,13 @@ func NewMerkleEyesApp(dbName string, cacheSize int) *MerkleEyesApp {
 	tree := iavl.NewIAVLTree(cacheSize, db)
 
 	if empty {
-		fmt.Println("no existing db, creating new db")
+		logger.Info("no existing db, creating new db")
 		db.Set(eyesStateKey, wire.BinaryBytes(MerkleEyesState{
 			Hash:   tree.Save(),
 			Height: initialHeight,
 		}))
 	} else {
-		fmt.Println("loading existing db")
+		logger.Info("loading existing db")
 	}
 
 	// Load merkle state
@@ -86,8 +114,9 @@ func NewMerkleEyesApp(dbName string, cacheSize int) *MerkleEyesApp {
 	var eyesState MerkleEyesState
 	err := wire.ReadBinaryBytes(eyesStateBytes, &eyesState)
 	if err != nil {
-		fmt.Println("error reading MerkleEyesState")
-		panic(err.Error())
+		logger.Error("error reading MerkleEyesState", "err", err)
+		// TODO: this should return an error, huh?
+		panic(err)
 	}
 
 	tree.Load(eyesState.Hash)
@@ -110,7 +139,7 @@ func (app *MerkleEyesApp) CloseDB() {
 // Info implements abci.Application. It returns the height, hash and size (in the data).
 // The height is the block that holds the transactions, not the apphash itself.
 func (app *MerkleEyesApp) Info() abci.ResponseInfo {
-	fmt.Printf("Info height %d hash %x\n", app.height, app.hash)
+	logger.Info("Info synced", "height", app.height, "hash", fmt.Sprintf("%X", app.hash))
 	return abci.ResponseInfo{
 		Data:             cmn.Fmt("size:%v", app.state.Committed().Size()),
 		LastBlockHeight:  app.height - 1,
@@ -158,7 +187,6 @@ func (app *MerkleEyesApp) doTx(tree merkle.Tree, tx []byte) abci.Result {
 		}
 
 		tree.Set(key, value)
-		// fmt.Println("SET", cmn.Fmt("%X", key), cmn.Fmt("%X", value))
 	case WriteRem: // Remove
 		key, n, err := wire.GetByteSlice(tx)
 		if err != nil {
@@ -180,7 +208,7 @@ func (app *MerkleEyesApp) Commit() abci.Result {
 
 	app.hash = app.state.Hash()
 	app.height++
-	fmt.Printf("height=%d,hash=%x\n", app.height, app.hash)
+	logger.Debug("Commit synced", "height", app.height, "hash", fmt.Sprintf("%X", app.hash))
 
 	if app.db != nil {
 		// This is in the batch with the Save, but not in the tree